+++ src/main/java/com/club_board/club_board_server/service/mail/EmailService.java
@@ -21,11 +21,23 @@ public class EmailService {
     @Value("${email.sender}")
     private String senderEmail;
 
-    private static final String AUTH_CODE_EMAIL_SUBJECT = "[CHIP_SAT] 인증번호 발송";
-    private static final String AUTH_CODE_EMAIL_BODY = """
+    private static final String PREFIX_EMAIL_SUBJECT = "[CHIP_SAT] ";
+
+    private static final String AUTH_CODE_EMAIL_SUBJECT = PREFIX_EMAIL_SUBJECT + "인증번호 발송";
+    private static final String AUTH_CODE_EMAIL_BODY =
+            """
             <h3>요청하신 인증 번호입니다.</h3>
             <h1>%s</h1>
-            <h3>감사합니다.<h3>
+            <h3>감사합니다.</h3>
+            """;
+
+    public static final String TEMP_PASSWORD_EMAIL_SUBJECT = PREFIX_EMAIL_SUBJECT + "임시 비밀번호 안내";
+    public static final String TEMP_PASSWORD_EMAIL_BODY =
+            """
+            <h3>요청하신 임시 비밀번호입니다.</h3>
+            <h1>%s</h1>
+            <h3>보안을 위해 로그인 후 반드시 비밀번호를 변경해 주세요!</h3>
+            <h3>감사합니다.</h3>
             """;
 
     @Async
@@ -37,6 +49,21 @@ public class EmailService {
             message.setSubject(AUTH_CODE_EMAIL_SUBJECT);
             String body = String.format(AUTH_CODE_EMAIL_BODY, number);
             message.setText(body,"UTF-8", "html");
+            javaMailSender.send(message);
+        } catch (MessagingException e) {
+            throw new BusinessException(ExceptionType.EMAIL_SEND_ERROR);
+        }
+    }
+
+    @Async
+    public void sendPasswordMail(String username, String password) {
+        MimeMessage message = javaMailSender.createMimeMessage();
+        try {
+            message.setFrom(senderEmail);
+            message.setRecipients(MimeMessage.RecipientType.TO, username);
+            message.setSubject(TEMP_PASSWORD_EMAIL_SUBJECT);
+            String body = String.format(TEMP_PASSWORD_EMAIL_BODY, password);
+            message.setText(body, "UTF-8", "html");
             javaMailSender.send(message);
         } catch (MessagingException e) {
             throw new BusinessException(ExceptionType.EMAIL_SEND_ERROR);
